Restructure website to support multiple games

+++ server/models/Score.py
@@ -6,5 +6,5 @@ from ..db import db
 class Score(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(256), nullable=False)
-    score = db.Column(db.integer, nullable=False)
+    score = db.Column(db.Float, nullable=False)
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
